+++ ConsoleUI/DependencyScreen.cs
@@ -2,6 +2,11 @@ using System;
 using System.Linq;
 using System.ComponentModel;
 using System.Collections.Generic;
+
+using CKAN.Versioning;
+#if NETFRAMEWORK
+using CKAN.Extensions;
+#endif
 using CKAN.ConsoleUI.Toolkit;
 
 namespace CKAN.ConsoleUI {
@@ -16,31 +21,26 @@ namespace CKAN.ConsoleUI {
         /// Initialize the screen
         /// </summary>
         /// <param name="mgr">Game instance manager containing instances</param>
-        /// <param name="registry">Registry of the current instance for finding mods</param>
+        /// <param name="reg">Registry of the current instance for finding mods</param>
         /// <param name="cp">Plan of mods to add and remove</param>
         /// <param name="rej">Mods that the user saw and did not select, in this pass or a previous pass</param>
         /// <param name="dbg">True if debug options should be available, false otherwise</param>
-        public DependencyScreen(GameInstanceManager mgr, Registry registry, ChangePlan cp, HashSet<string> rej, bool dbg) : base()
+        public DependencyScreen(GameInstanceManager mgr, Registry reg, ChangePlan cp, HashSet<string> rej, bool dbg) : base()
         {
-            debug     = dbg;
-            manager   = mgr;
-            plan      = cp;
-            this.registry = registry;
-            installer = new ModuleInstaller(manager.CurrentInstance, manager.Cache, this);
-            rejected  = rej;
+            debug    = dbg;
+            manager  = mgr;
+            plan     = cp;
+            registry = reg;
+            rejected = rej;
 
-            AddObject(new ConsoleLabel(
-                1, 2, -1,
-                () => Properties.Resources.RecommendationsLabel
-            ));
+            AddObject(new ConsoleLabel(1, 2, -1,
+                                       () => Properties.Resources.RecommendationsLabel));
 
-            HashSet<CkanModule> sourceModules = new HashSet<CkanModule>();
-            sourceModules.UnionWith(plan.Install);
-            sourceModules.UnionWith(new HashSet<CkanModule>(
-                ReplacementIdentifiers(plan.Replace)
-                    .Select(id => registry.InstalledModule(id).Module)
-            ));
-            generateList(sourceModules);
+            generateList(new ModuleInstaller(manager.CurrentInstance, manager.Cache, this),
+                         plan.Install
+                             .Concat(ReplacementModules(plan.Replace,
+                                                        manager.CurrentInstance.VersionCriteria()))
+                             .ToHashSet());
 
             dependencyList = new ConsoleListBox<Dependency>(
                 1, 4, -1, -2,
@@ -49,32 +49,37 @@ namespace CKAN.ConsoleUI {
                     new ConsoleListBoxColumn<Dependency>() {
                         Header   = Properties.Resources.RecommendationsInstallHeader,
                         Width    = 7,
-                        Renderer = (Dependency d) => StatusSymbol(d.module)
+                        Renderer = (Dependency d) => StatusSymbol(d.module),
                     },
                     new ConsoleListBoxColumn<Dependency>() {
                         Header   = Properties.Resources.RecommendationsNameHeader,
                         Width    = null,
-                        Renderer = (Dependency d) => d.module.ToString()
+                        Renderer = (Dependency d) => d.module.ToString(),
                     },
                     new ConsoleListBoxColumn<Dependency>() {
                         Header   = Properties.Resources.RecommendationsSourcesHeader,
                         Width    = 42,
-                        Renderer = (Dependency d) => string.Join(", ", d.dependents)
+                        Renderer = (Dependency d) => string.Join(", ", d.dependents),
                     }
                 },
                 1, 0, ListSortDirection.Descending
             );
             dependencyList.AddTip("+", Properties.Resources.Toggle);
             dependencyList.AddBinding(Keys.Plus, (object sender, ConsoleTheme theme) => {
-                ChangePlan.toggleContains(accepted, dependencyList.Selection.module);
+                var mod = dependencyList.Selection.module;
+                if (accepted.Contains(mod) || TryWithoutConflicts(accepted.Append(mod))) {
+                    ChangePlan.toggleContains(accepted, mod);
+                }
                 return true;
             });
 
             dependencyList.AddTip($"{Properties.Resources.Ctrl}+A", Properties.Resources.SelectAll);
             dependencyList.AddBinding(Keys.CtrlA, (object sender, ConsoleTheme theme) => {
-                foreach (var kvp in dependencies) {
-                    if (!accepted.Contains(kvp.Key)) {
-                        ChangePlan.toggleContains(accepted, kvp.Key);
+                if (TryWithoutConflicts(dependencies.Keys)) {
+                    foreach (var kvp in dependencies) {
+                        if (!accepted.Contains(kvp.Key)) {
+                            ChangePlan.toggleContains(accepted, kvp.Key);
+                        }
                     }
                 }
                 return true;
@@ -89,10 +94,9 @@ namespace CKAN.ConsoleUI {
             dependencyList.AddTip(Properties.Resources.Enter, Properties.Resources.Details);
             dependencyList.AddBinding(Keys.Enter, (object sender, ConsoleTheme theme) => {
                 if (dependencyList.Selection != null) {
-                    LaunchSubScreen(theme, new ModInfoScreen(
-                        manager, registry, plan,
-                        dependencyList.Selection.module,
-                        debug));
+                    LaunchSubScreen(theme, new ModInfoScreen(manager, reg, plan,
+                                                             dependencyList.Selection.module,
+                                                             debug));
                 }
                 return true;
             });
@@ -102,53 +106,41 @@ namespace CKAN.ConsoleUI {
             AddTip(Properties.Resources.Esc, Properties.Resources.Cancel);
             AddBinding(Keys.Escape, (object sender, ConsoleTheme theme) => {
                 // Add everything to rejected
-                foreach (var kvp in dependencies) {
-                    rejected.Add(kvp.Key.identifier);
-                }
+                rejected.UnionWith(dependencies.Keys.Select(m => m.identifier));
                 return false;
             });
 
             AddTip("F9", Properties.Resources.Accept);
             AddBinding(Keys.F9, (object sender, ConsoleTheme theme) => {
-                foreach (CkanModule mod in accepted) {
-                    plan.Install.Add(mod);
+                if (TryWithoutConflicts(accepted)) {
+                    plan.Install.UnionWith(accepted);
+                    // Add the rest to rejected
+                    rejected.UnionWith(dependencies.Keys
+                                                   .Except(accepted)
+                                                   .Select(m => m.identifier));
+                    return false;
                 }
-                // Add the rest to rejected
-                foreach (var kvp in dependencies) {
-                    if (!accepted.Contains(kvp.Key)) {
-                        rejected.Add(kvp.Key.identifier);
-                    }
-                }
-                return false;
+                return true;
             });
         }
 
         /// <summary>
         /// Put CKAN 1.25.5 in top left corner
         /// </summary>
-        protected override string LeftHeader()
-        {
-            return $"CKAN {Meta.GetVersion()}";
-        }
+        protected override string LeftHeader() => $"{Meta.GetProductName()} {Meta.GetVersion()}";
 
         /// <summary>
         /// Put description in top center
         /// </summary>
-        protected override string CenterHeader()
-        {
-            return Properties.Resources.RecommendationsTitle;
-        }
+        protected override string CenterHeader() => Properties.Resources.RecommendationsTitle;
 
         /// <summary>
         /// Return whether there are any options to show.
         /// ModListScreen uses this to avoid showing this screen when empty.
         /// </summary>
-        public bool HaveOptions()
-        {
-            return dependencies.Count > 0;
-        }
+        public bool HaveOptions() => dependencies.Count > 0;
 
-        private void generateList(HashSet<CkanModule> inst)
+        private void generateList(ModuleInstaller installer, HashSet<CkanModule> inst)
         {
             if (installer.FindRecommendations(
                 inst, new List<CkanModule>(inst), registry as Registry,
@@ -156,50 +148,68 @@ namespace CKAN.ConsoleUI {
                 out Dictionary<CkanModule, List<string>> suggestions,
                 out Dictionary<CkanModule, HashSet<string>> supporters
             )) {
-                foreach (var kvp in recommendations) {
-                    dependencies.Add(kvp.Key, new Dependency() {
-                        module         = kvp.Key,
-                        defaultInstall = kvp.Value.Item1,
-                        dependents     = kvp.Value.Item2.OrderBy(d => d).ToList()
-                    });
-                    if (kvp.Value.Item1) {
-                        accepted.Add(kvp.Key);
-                    }
-                }
-                foreach (var kvp in suggestions) {
-                    dependencies.Add(kvp.Key, new Dependency() {
-                        module         = kvp.Key,
-                        defaultInstall = false,
-                        dependents     = kvp.Value.OrderBy(d => d).ToList()
+                foreach ((CkanModule mod, Tuple<bool, List<string>> checkedAndDependents) in recommendations) {
+                    dependencies.Add(mod, new Dependency() {
+                        module     = mod,
+                        dependents = checkedAndDependents.Item2.OrderBy(d => d).ToList()
                     });
                 }
-                foreach (var kvp in supporters) {
-                    dependencies.Add(kvp.Key, new Dependency() {
-                        module         = kvp.Key,
-                        defaultInstall = false,
-                        dependents     = kvp.Value.OrderBy(d => d).ToList()
+                foreach ((CkanModule mod, List<string> dependents) in suggestions) {
+                    dependencies.Add(mod, new Dependency() {
+                        module     = mod,
+                        dependents = dependents.OrderBy(d => d).ToList()
                     });
                 }
+                foreach ((CkanModule mod, HashSet<string> dependents) in supporters) {
+                    dependencies.Add(mod, new Dependency() {
+                        module     = mod,
+                        dependents = dependents.OrderBy(d => d).ToList()
+                    });
+                }
+                // Check the default checkboxes
+                accepted.UnionWith(recommendations.Where(kvp => kvp.Value.Item1)
+                                                  .Select(kvp => kvp.Key));
             }
         }
 
-        private IEnumerable<string> ReplacementIdentifiers(IEnumerable<string> replaced_identifiers)
-        {
-            foreach (string replaced in replaced_identifiers) {
-                ModuleReplacement repl = registry.GetReplacement(
-                    replaced, manager.CurrentInstance.VersionCriteria()
-                );
-                if (repl != null) {
-                    yield return repl.ReplaceWith.identifier;
-                }
-            }
-        }
+        private IEnumerable<CkanModule> ReplacementModules(IEnumerable<string> replaced_identifiers,
+                                                           GameVersionCriteria crit)
+            => replaced_identifiers.Select(replaced => registry.GetReplacement(replaced, crit))
+                                   .Where(repl => repl != null)
+                                   .Select(repl => repl.ReplaceWith);
 
         private string StatusSymbol(CkanModule mod)
+            => accepted.Contains(mod) ? installing
+                                      : notinstalled;
+
+        private bool TryWithoutConflicts(IEnumerable<CkanModule> toAdd)
         {
-            return accepted.Contains(mod)
-                ? installing
-                : notinstalled;
+            if (HasConflicts(toAdd, out List<string> conflictDescriptions)) {
+                RaiseError("{0}", string.Join(Environment.NewLine,
+                                              conflictDescriptions));
+                return false;
+            }
+            return true;
+        }
+
+        private bool HasConflicts(IEnumerable<CkanModule> toAdd,
+                                  out List<string>        descriptions)
+        {
+            try
+            {
+                var resolver = new RelationshipResolver(
+                    plan.Install.Concat(toAdd).Distinct(),
+                    plan.Remove.Select(ident => registry.InstalledModule(ident)?.Module),
+                    RelationshipResolverOptions.ConflictsOpts(), registry,
+                    manager.CurrentInstance.VersionCriteria());
+                descriptions = resolver.ConflictDescriptions.ToList();
+                return descriptions.Count > 0;
+            }
+            catch (DependencyNotSatisfiedKraken k)
+            {
+                descriptions = new List<string>() { k.Message };
+                return true;
+            }
         }
 
         private readonly HashSet<CkanModule> accepted = new HashSet<CkanModule>();
@@ -207,15 +217,14 @@ namespace CKAN.ConsoleUI {
 
         private readonly IRegistryQuerier    registry;
         private readonly GameInstanceManager manager;
-        private readonly ModuleInstaller     installer;
         private readonly ChangePlan          plan;
         private readonly bool                debug;
 
         private readonly Dictionary<CkanModule, Dependency> dependencies = new Dictionary<CkanModule, Dependency>();
         private readonly ConsoleListBox<Dependency>         dependencyList;
 
-        private static readonly string notinstalled = " ";
-        private static readonly string installing   = "+";
+        private const string notinstalled = " ";
+        private const string installing   = "+";
     }
 
     /// <summary>
@@ -224,14 +233,9 @@ namespace CKAN.ConsoleUI {
     public class Dependency {
 
         /// <summary>
-        /// Identifier of mod
+        /// The mod
         /// </summary>
-        public CkanModule   module;
-
-        /// <summary>
-        /// True if we default to installing, false otherwise
-        /// </summary>
-        public bool         defaultInstall;
+        public CkanModule module;
 
         /// <summary>
         /// List of mods that recommended or suggested this mod